new file for easingFunctions
